rewrote modules to be called from main_gui
