+++ GUI/modules/Communicate.py
@@ -1,0 +1,7 @@
+from PyQt5.QtCore import pyqtSignal, QObject
+def initializeGlobalVars():
+    global runningTest
+    runningTest = False
+
+class Communicate(QObject):
+    GUI_signal = pyqtSignal(str)
